[REFACTOR] 예외처리 기능 패키지 재분류
